Use Express Router in the admin-api.
