+++ services/admin-api/src/middlewares/index.js
@@ -1,0 +1,5 @@
+const magic = require("./magic");
+
+module.exports = {
+  magic,
+};
