+++ services/admin-api/src/routes/index.js
@@ -1,0 +1,7 @@
+const user = require("./user");
+const website = require("./website");
+
+module.exports = {
+  user,
+  website,
+};
